Move service into subdirectory.
