+++ src/app/playback-control/playback-control.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {AudioPlayerService} from "../services/audio-player.service";
+import {AudioPlayerService} from "../services/audio-player/audio-player.service";
 
 @Component({
   selector: 'app-playback-control',
